Refactored Enemy_Controller Script
Updated and refactored the Enemy_Controller script to be more performant by removed methods from the update method and streamlining logic within the script. Updated Particle Effects to auto-destroy to increase performance. Removed One of the Traps from each map and updated the boulders to be based on a prefab